refactor: 승리자를 출력하는 코드 포매터 개선
